+++ src/App.tsx
@@ -76,7 +76,7 @@ const App = () => {
 				}
 
 				<div className="p-1 pt-0 w-100">
-					<AppBreadcrumb className="mt-1" />
+					<AppBreadcrumb className="mt-1 mb-1" />
 
 					<Switch>
 						{appRoutes &&